requirements.txt
# data libraries
numpy
pandas
joblib

# helpers
absl-py

# slurm-utils itself
slurm-utils